Refactor - Create file for BaseChecker
It was 'hidden' in checkers.__init__.py
